requirements.txt
aiogram==2.25.2
python-dotenv
openai